Read and analyze univariate time series data for EU datathon

Extract and analyze univariate time series data for EU datathon

predict_onrg(json_data, time="years", amount="values", order=NULL,
 prediction_steps= 5, tojson= TRUE)

Arguments

json_data
The json text or link with the EU Energy data frame or matrix

time
Time label of the json time series data

amount
Amount label of the json time series data

order
An integer vector of length 3 specifying the order of the Arima model

prediction_steps
The number of prediction steps

tojson
If FALSE returns a list object

Value

A json string with the resulted parameters of the ts.analysis function.

Details

This function is used to predict eu energy statistics/market data in OpEnergy application.